feat: harden contracts with constraint handling, nullable stats, and test coverage
Add DB constraint-violation translation in JdbcSleepLogRepository to map
PostgreSQL check/exclusion violations to domain exceptions. Make SleepStats
averages nullable for empty periods. Add UserRepository.saveUser with JDBC
implementation. Tighten OpenAPI spec (nullable averages, userId minimum,
clarified date ranges). Fix build.gradle dependency setup and update CLAUDE.md.

Refs: #4

+++ sleep/src/main/kotlin/com/noom/interview/fullstack/sleep/model/SleepLog.kt
@@ -23,5 +23,5 @@ data class SleepLog(
     val wakeTimeZone: ZoneId
 ) {
     val durationMinutes: Int
-        get() = Duration.between(bedTime, wakeTime).toMinutes().toInt()
+        get() = Math.toIntExact(Duration.between(bedTime, wakeTime).toMinutes())
 }
